+++ src/components/AIReview/AiSummaryTable.styl
@@ -1,0 +1,115 @@
+.ai-summary-container {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+
+    .ai-summary-table {
+        th, td {
+            text-align: center;
+        }
+
+        .move-count-container {
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            position: relative;
+
+            .move-list-button {
+                background: none;
+                border: none;
+                color: #666;
+                cursor: pointer;
+                font-size: 0.8em;
+                opacity: 0.7;
+                transition: all 0.2s;
+                display: flex;
+                align-items: center;
+                justify-content: center;
+                width: 20px;
+                height: 20px;
+                border-radius: 3px;
+
+                &.invisible {
+                    opacity: 0;
+                    cursor: default;
+                    pointer-events: none;
+                }
+
+                &:hover {
+                    opacity: 1;
+                    background-color: #f0f0f0;
+                    transform: translateY(1px);
+                }
+            }
+        }
+    }
+}
+
+.category-move-popover {
+    position: absolute;
+    top: 100%;
+    left: 50%;
+    transform: translateX(-50%);
+    box-shadow: 0px 2px 3px rgba(0, 0, 0, 0.2);
+    border: 1px solid transparent;
+    box-shadow: 0px 2px 10px 0px rgba(0, 0, 0, 0.16), 0px 2px 5px 0px rgba(0, 0, 0, 0.15);
+    themed: background-color shade4;
+    themed: border-color shade3;
+    themed: color fg;
+    z-index: z.popover;
+    min-width: 200px;
+    max-width: 300px;
+    max-height: 300px;
+    overflow-y: auto;
+
+    .category-move-header {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        padding: 8px 12px;
+        border-bottom: 1px solid transparent;
+        themed: border-bottom-color shade3;
+        font-weight: bold;
+
+        .close-button {
+            background: none;
+            border: none;
+            font-size: 1.2em;
+            cursor: pointer;
+            padding: 0 4px;
+            themed: color shade1;
+
+            &:hover {
+                themed: color fg;
+            }
+        }
+    }
+
+    .category-move-content {
+        padding: 12px;
+
+        .move-numbers {
+            display: flex;
+            flex-wrap: wrap;
+            gap: 8px;
+
+            .move-number {
+                themed: background-color shade3;
+                padding: 4px 8px;
+                border-radius: 3px;
+                font-size: 0.9em;
+                cursor: pointer;
+
+                &:hover {
+                    themed: background-color shade2;
+                }
+            }
+        }
+
+        .no-moves {
+            themed: color shade1;
+            font-style: italic;
+            text-align: center;
+        }
+    }
+}
